Arreglo de errores y ultimos retoques

+++ views/users/view.php
@@ -38,28 +38,30 @@ $this->title = $model->username . "'s profile";
         ]) ?>
         <div class="row">
             <?= Html::a('Update profile', ['update', 'id' => $model->Id], ['class' => 'btn btn-success'])?> 
-            <?php if(Yii::$app->user->identity->role == "Admin" || Yii::$app->user->identity->role == "Owner"):?>
-                <?= Html::a('Unban user', ['unban', 'id' => $model->Id], ['class' => 'btn btn-info'])?> 
-                <?= Html::a('Ban user', ['ban', 'id' => $model->Id], ['class' => 'btn btn-danger',
-                'data' => ['confirm' => 'Are you sure you want to ban the user?','method'=>'post']])?> 
-                <?php $form = ActiveForm::begin(); ?>
-                    <?php Modal::begin([
-                        'title' => 'Select role',
-                        'toggleButton' => ['label' => 'Change role', 'class' => 'btn btn-primary'],
-                    ]);
+            <?php if(!(Yii::$app->user->isGuest)):?>
+                <?php if(Yii::$app->user->identity->role == "Admin" || Yii::$app->user->identity->role == "Owner"):?>
+                    <?= Html::a('Unban user', ['unban', 'id' => $model->Id], ['class' => 'btn btn-info'])?> 
+                    <?= Html::a('Ban user', ['ban', 'id' => $model->Id], ['class' => 'btn btn-danger',
+                    'data' => ['confirm' => 'Are you sure you want to ban the user?','method'=>'post']])?> 
+                    <?php $form = ActiveForm::begin(); ?>
+                        <?php Modal::begin([
+                            'title' => 'Select role',
+                            'toggleButton' => ['label' => 'Change role', 'class' => 'btn btn-primary'],
+                        ]);
 
-                    echo  $form->field($model, 'role')->label('Role')->dropDownList(
-                        ['User' =>'User',
-                        'Editor' =>'Editor',
-                        'Moderator' =>'Moderator',
-                        'Admin' =>'Admin'],
-                        ['selected' => $model->role]
-                    );
+                        echo  $form->field($model, 'role')->label('Role')->dropDownList(
+                            ['User' =>'User',
+                            'Editor' =>'Editor',
+                            'Moderator' =>'Moderator',
+                            'Admin' =>'Admin'],
+                            ['selected' => $model->role]
+                        );
 
-                    echo Html::submitButton('Save', ['class' => 'btn btn-success']);
+                        echo Html::submitButton('Save', ['class' => 'btn btn-success']);
 
-                    Modal::end();?>
-                <?php ActiveForm::end(); ?>
+                        Modal::end();?>
+                    <?php ActiveForm::end(); ?>
+                <?php endif ?>
             <?php endif ?>
         </div>
